Ajout du Makefile fonctionnel mon gars

+++ ft_calloc.c
@@ -15,7 +15,7 @@ void	*calloc(size_t count, size_t size)
 {
 	void	*s;
 
-	s = (size_t)malloc(size * count);
+	s = malloc(size * count);
 	if (!s)
 		return (NULL);
 	ft_bzero(s, size * count);
